WIP loplugin:droplong in tools/gen.hxx and vcl
large patch because the gen.hxx stuff just keeps propogating
everywhere.

+++ sc/qa/unit/bugfix-test.cxx
@@ -160,8 +160,8 @@ void ScFiltersTest::testTdf91979()
     Point aPos = aViewData.GetScrPos(MAXCOL - 1, 10000, SC_SPLIT_TOPLEFT, true);
     int nColWidth = ScViewData::ToPixel(pDoc->GetColWidth(0, 0), aViewData.GetPPTX());
     int nRowHeight = ScViewData::ToPixel(pDoc->GetRowHeight(0, 0), aViewData.GetPPTY());
-    CPPUNIT_ASSERT_EQUAL(static_cast<long>((MAXCOL - 1) * nColWidth), aPos.getX());
-    CPPUNIT_ASSERT_EQUAL(static_cast<long>(10000 * nRowHeight), aPos.getY());
+    CPPUNIT_ASSERT_EQUAL((MAXCOL - 1) * nColWidth, aPos.getX());
+    CPPUNIT_ASSERT_EQUAL(10000 * nRowHeight, aPos.getY());
 
     xComponent->dispose();
 }
