+++ sw/source/core/doc/notxtfrm.cxx
@@ -381,7 +381,7 @@ void SwNoTextFrame::GetGrfArea( SwRect &rRect, SwRect* pOrigRect ) const
     else
     {
         nLeftCrop = std::max( aOrigSz.Width() -
-                            (rCrop.GetRight() + rCrop.GetLeft()), long(1) );
+                            (rCrop.GetRight() + rCrop.GetLeft()), 1 );
         const double nScale = double(Prt().Width())  / double(nLeftCrop);
         nLeftCrop  = long(nScale * -rCrop.GetLeft() );
         nRightCrop = long(nScale * -rCrop.GetRight() );
@@ -403,7 +403,7 @@ void SwNoTextFrame::GetGrfArea( SwRect &rRect, SwRect* pOrigRect ) const
     }
     else
     {
-        nTopCrop = std::max( aOrigSz.Height() - (rCrop.GetTop() + rCrop.GetBottom()), long(1) );
+        nTopCrop = std::max( aOrigSz.Height() - (rCrop.GetTop() + rCrop.GetBottom()), 1 );
         const double nScale = double(Prt().Height()) / double(nTopCrop);
         nTopCrop   = long(nScale * -rCrop.GetTop() );
         nBottomCrop= long(nScale * -rCrop.GetBottom() );
